+++ frontend/store/authentication.js
@@ -1,4 +1,4 @@
-import Cookies from "~/utils/Cookie"
+import Cookies from "~/utils/Cookies"
 import User from "~/types/User"
 const cookies = new Cookies()
